+++ packages/snap-preact-components/src/components/Atoms/Image/Image.test.tsx
@@ -12,22 +12,11 @@ describe('image Component', () => {
 
 	it('renders', () => {
 		const rendered = render(<Image alt={result.name} src={result.thumbnailImageUrl} />);
-		const imageElement = rendered.container.querySelector('.ss__image');
+		const imageElement = rendered.container.querySelector('.ss__image img');
 		expect(imageElement).toBeInTheDocument();
 		expect(imageElement).toHaveAttribute('src', result.thumbnailImageUrl);
 		expect(imageElement).toHaveAttribute('alt', result.name);
 		expect(imageElement).toHaveAttribute('title', result.name);
-
-		expect(imageElement.classList).toHaveLength(2);
-	});
-
-	it('renders with classname', () => {
-		const className = 'classy';
-		const rendered = render(<Image className={className} alt={result.name} src={result.thumbnailImageUrl} />);
-
-		const imageElement = rendered.container.querySelector('.ss__image');
-
-		expect(imageElement).toHaveClass(className);
 	});
 
 	it('renders with classname', () => {
@@ -48,7 +37,7 @@ describe('image Component', () => {
 	describe('Working Image', () => {
 		it('renders image', () => {
 			const rendered = render(<Image alt={result.name} src={result.thumbnailImageUrl} />);
-			const imageElement = rendered.container.querySelector('.ss__image');
+			const imageElement = rendered.container.querySelector('.ss__image img');
 			expect(imageElement).toBeInTheDocument();
 			expect(imageElement).toHaveAttribute('src', result.thumbnailImageUrl);
 		});
@@ -57,14 +46,14 @@ describe('image Component', () => {
 	describe('Broken Image', () => {
 		it('should display default fallback image', () => {
 			const rendered = render(<Image alt={badResult.name} src={badResult.thumbnailImageUrl} />);
-			const imageElement = rendered.container.querySelector('.ss__image');
+			const imageElement = rendered.container.querySelector('.ss__image img');
 			expect(imageElement).toHaveAttribute('src', FALLBACK_IMAGE_URL);
 		});
 
 		it('should display custom fallback image', () => {
 			const fallbackImage = 'https://www.telegraph.co.uk/content/dam/Pets/spark/royal-canin/happy-puppy-xlarge.jpg?imwidth=1200';
 			const rendered = render(<Image alt={badResult.name} src={badResult.thumbnailImageUrl} fallback={fallbackImage} />);
-			const imageElement = rendered.container.querySelector('.ss__image');
+			const imageElement = rendered.container.querySelector('.ss__image img');
 			expect(imageElement).toHaveAttribute('src', fallbackImage);
 		});
 	});