+++ packages/snap-preact-components/src/components/Molecules/Result/Result.test.tsx
@@ -14,25 +14,25 @@ describe('Result Component', () => {
 
 	it('renders image', () => {
 		const rendered = render(<Result result={searchResponse.results[0]} />);
-		const imageElement = rendered.container.querySelector('.ss__result .ss__result__wrapper__image .ss__image');
+		const imageElement = rendered.container.querySelector('.ss__result .ss__result__image-wrapper .ss__image img');
 		expect(imageElement).toBeInTheDocument();
 	});
 
 	it('renders badge', () => {
 		const rendered = render(<Result result={searchResponse.results[0]} />);
-		const badgeElement = rendered.container.querySelector('.ss__result .ss__result__wrapper__image .ss__badge');
+		const badgeElement = rendered.container.querySelector('.ss__result .ss__result__image-wrapper .ss__badge');
 		expect(badgeElement).toBeInTheDocument();
 	});
 
 	it('renders title', () => {
 		const rendered = render(<Result result={searchResponse.results[0]} />);
-		const title = rendered.container.querySelector('.ss__result .ss__result__wrapper__details .ss__result__wrapper__details__title').textContent;
+		const title = rendered.container.querySelector('.ss__result .ss__result__details .ss__result__details__title').textContent;
 		expect(title).toBe(searchResponse.results[0].mappings.core.name);
 	});
 
 	it('renders pricing', () => {
 		const rendered = render(<Result result={searchResponse.results[0]} />);
-		const priceElement = rendered.container.querySelectorAll('.ss__result .ss__result__wrapper__details__pricing .ss__price');
+		const priceElement = rendered.container.querySelectorAll('.ss__result .ss__result__details__pricing .ss__price');
 		expect(priceElement[0]).toBeInTheDocument();
 		expect(priceElement.length).toBe(2);
 	});
@@ -43,7 +43,7 @@ describe('Result Component', () => {
 			detailSlot: <div className="details">Add to cart'</div>,
 		};
 		const rendered = render(<Result {...args} />);
-		const detailsElement = rendered.container.querySelector('.ss__result .ss__result__wrapper__details .details');
+		const detailsElement = rendered.container.querySelector('.ss__result .ss__result__details .details');
 		expect(detailsElement).toBeInTheDocument();
 		expect(detailsElement).toHaveTextContent('Add to cart');
 	});
@@ -56,9 +56,9 @@ describe('Result Component', () => {
 			hidePricing: true,
 		};
 		const rendered = render(<Result {...args} />);
-		const badgeElement = rendered.container.querySelector('.ss__result .ss__result__wrapper__image .ss__badge');
-		const titleElement = rendered.container.querySelector('.ss__result .ss__result__wrapper__details .ss__result__wrapper__details__title');
-		const priceElement = rendered.container.querySelector('.ss__result .ss__result__wrapper__details__pricing .ss__price');
+		const badgeElement = rendered.container.querySelector('.ss__result .ss__result__image-wrapper .ss__badge');
+		const titleElement = rendered.container.querySelector('.ss__result .ss__result__details .ss__result__wrapper__details__title');
+		const priceElement = rendered.container.querySelector('.ss__result .ss__result__details__pricing .ss__price');
 		expect(badgeElement).not.toBeInTheDocument();
 		expect(titleElement).not.toBeInTheDocument();
 		expect(priceElement).not.toBeInTheDocument();
@@ -66,7 +66,7 @@ describe('Result Component', () => {
 
 	it('should display a fallback image', () => {
 		const rendered = render(<Result result={searchResponse.results[1]} />);
-		const imageElement = rendered.container.querySelector('.ss__result .ss__result__wrapper__image .ss__image');
+		const imageElement = rendered.container.querySelector('.ss__result .ss__result__image-wrapper .ss__image img');
 		expect(imageElement).toHaveAttribute('src', FALLBACK_IMAGE_URL);
 	});
 });
@@ -86,7 +86,7 @@ describe('Result theming works', () => {
 			</ThemeProvider>
 		);
 		const result = rendered.container.querySelector('.ss__result');
-		const title = rendered.container.querySelector('.ss__result__wrapper__details__title');
+		const title = rendered.container.querySelector('.ss__result__details__title');
 		expect(result).toBeInTheDocument();
 		expect(title).not.toBeInTheDocument();
 	});
@@ -101,7 +101,7 @@ describe('Result theming works', () => {
 		};
 		const rendered = render(<Result result={searchResponse.results[0]} theme={propTheme} />);
 		const result = rendered.container.querySelector('.ss__result');
-		const title = rendered.container.querySelector('.ss__result__wrapper__details__title');
+		const title = rendered.container.querySelector('.ss__result__details__title');
 		expect(result).toBeInTheDocument();
 		expect(title).not.toBeInTheDocument();
 	});
@@ -130,7 +130,7 @@ describe('Result theming works', () => {
 
 		const badge = rendered.container.querySelector('.ss__badge');
 		const result = rendered.container.querySelector('.ss__result');
-		const title = rendered.container.querySelector('.ss__result__wrapper__details__title');
+		const title = rendered.container.querySelector('.ss__result__details__title');
 		expect(result).toBeInTheDocument();
 		expect(title).toBeInTheDocument();
 		expect(badge).not.toBeInTheDocument();