+++ packages/snap-preact-demo/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/searchspring/snap/compare/v0.4.0...v0.5.0) (2021-08-23)
+
+
+### Bug Fixes
+
+* **autocomplete:** add flex basis to facets, remove demo store autocomplete theme.results duplicate ([0b79a73](https://github.com/searchspring/snap/commit/0b79a739f1cdf86dda7e871ca3bd0f2881569c0f))
+
+
+
+
+
+# [0.4.0](https://github.com/searchspring/snap/compare/v0.3.47...v0.4.0) (2021-08-23)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## [0.3.47](https://github.com/searchspring/snap/compare/v0.3.46...v0.3.47) (2021-08-23)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## 0.3.46 (2021-08-21)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## 0.3.45 (2021-08-20)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## 0.3.44 (2021-08-19)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## 0.3.43 (2021-08-19)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
 ## 0.3.42 (2021-08-18)
 
 **Note:** Version bump only for package @searchspring/snap-preact-demo